% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFindCurve.R
\name{CleanFindCurve}
\alias{CleanFindCurve}
\title{Deletes on request all files produced by the FindCurve package.}
\usage{
CleanFindCurve(str = NULL)
}
\arguments{
\item{str}{Character (optional). Only valid argument is 'F'. If not or wrongly
specified the user will be asked whether to do a full clean up or whether to
quit the clean up.}
}
\value{
None.
}
\description{
\code{CleanFindCurve} deletes all FindCurve result files (default) and/or
all executables (hit 'F') in the current directory.
}
\examples{
CleanFindCurve()

CleanFindCurve("F")

}
