% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showsystem.R
\name{showsystem}
\alias{showsystem}
\title{Shows the system definition file of one of the examples provided with FindCurve}
\usage{
showsystem(modelname = NULL)
}
\arguments{
\item{modelname}{(string)
\preformatted{}
              Name of the example system to be displayed.}
}
\description{
\code{showsystem} displays the file contents of one of the systems of equations
that is provided as an example.
}
\details{
showsystem(modelname = NULL)
}
\examples{
showsystem("CjaP.h")

}
